resty/http/proxy: fix incorrect connection reuse
Fixes https://issues.jboss.org/browse/THREESCALE-2205

By default, the 'connect' method of lua-resty-http, uses the ip and the
host as the key to find a connection in the pool of idle connections.
Ref: https://github.com/ledgetech/lua-resty-http#connect

That can cause problems in the case of HTTPS requests as mentioned in
the issue linked above. For example, when 2 different domains resolve to
the same IP, the connection is reused, and it should not be.

The 'connect_proxy' method of the modified module was using correct IDs
for the pool, but the 'connect_direct' method was not.