t/apicast-policy-3scale-batcher: change test to make sure it uses the caching policy
Unfortunately the test did not properly check that the batching policy
works well with the caching one.

The reason is that it defined a backend that returns a limits exceeded
error and then a 5xx in all the rest. Then, the test expected to see a
"limits exceeded" in the first response, and then, an authorization
denied in the rest. The problem is that that's exactly what's returned
without using the caching policy. The caching policy was not working
correctly and we did not catch that in the tests because of that.

Now, the backend returns 200 in the first request and then, 5xx in all
the rest. In order to check that the caching policy is working
correctly, we verify that the first 200 is cached and that all the
requests succeed even when backend returns error.