[Test] Fix double stub
Over the last days some issues has been triggered due the stub was not
reverted correctly. The revert happens on luaasert.state and it's called
on `spec/luaasert_spect.lua` but if a double stub happens, the first
stub is leaked and it's not reverted.

This fixes the cases where the double stub happens, in the near future a
new commit will be added to check that the stubs are cleaned correctly.

Signed-off-by: Eloy Coto <eloy.coto@gmail.com>