Routing: Fix query args on replace_path
Due to the changes made in the routing policy to be able to modify
request path the query args was appended, and in this case, unexcaped to
the upstream repo.

This change adds a resty/url_helper function; this code will be moved to
resty-url project in the following weeks.

Signed-off-by: Eloy Coto <eloy.coto@gmail.com>