Policy: Add GC policy for shared dicts.
This policy allows APICast to remove data that it's staled from shared
dictionaries.  By default, Openresty only deletes this when trying to
write a new entry in the shared dict.  When no new requests in this
instance, will keep allocated the memory forever.

The main reason for this change was a user using blue-green deployment,
where the proxy is not receiving more traffic. However, the shared
dictionaries are full, so the data is still allocated and cannot be used
by other processes.

Related to: https://issues.redhat.com/browse/THREESCALE-6718

Signed-off-by: Eloy Coto <eloy.coto@acalustra.com>