Keepalive: pool based on variable.
When using keepalive connections, can be problematic when the
microservice is behind a LB.

When the upstream is shared accross services, if a connection is like:
APICast->LB->serviceA, the pool will be based on the IP. For the
serviceB, the request will land to the serviceA, because no SNI check at
all.

When using plaintext HTTP, most of the LB check the HOST header to
verify the upstream Service, but Host header is not in there when the LB
only do passthrough.

Fix THREESCALE-6849

Signed-off-by: Eloy Coto <eloy.coto@acalustra.com>