Logs: disable invalid warning when no api_backend
Since API as a product, api_backend is no longer a required parameter of
the config. Due to the change, a warning message was in place on each
request due to invalid upstream.

With this change, in case of no api_backend, a nil object is returned
and no error message at all.

Fix THREESCALE-5225

Signed-off-by: Eloy Coto <eloy.coto@acalustra.com>